requirements.txt
torch>=2.6.0
torchvision>=0.21.0
numpy>=1.26

typer>=0.15.2
fastapi>=0.115.12
uvicorn>=0.34.2
wandb>=0.16
tqdm>=4.66

hydra-core>=1.3
omegaconf>=2.3

scikit-learn>=1.3
pillow>=10.0
matplotlib>=3.7

loguru>=0.7.3
